DeleteStudent (#6)

+++ LearningManagementSystem/Models/Student.cs
@@ -11,10 +11,10 @@ namespace LearningManagementSystem.Models
         public string FirstName { get; set; }
         public string LastName { get; set; }
         public string Program { get; set; }
-        public double Phone { get; set; }
+        public float Phone { get; set; }
         public string Email { get; set; }
 
-        public Student(int id, string firstName, string lastName, string program, double phone, string email)
+        public Student(int id, string firstName, string lastName, string program, float phone, string email)
         {
             Id = id;
             FirstName = firstName;
